Update `ref/` classes to match move to `net462`
